ref(starfish): Clean out initial Starfish scaffolding (#47282)
Setting up a better workspace for us. Three main changes:

1. Add subrouting. `/starfish`, `/starfish/api`, `/starfish/cache`,
`/starfish/database` have their own routes. `/starfish` still has the
original content, the others are dummy components with some basic
content and layout
2. Removed as much code as possible. I'm not done yet, but this removes
a lot of code. I changed the landing page to render a barebones graph,
using our event and graph primitives. This skips MEP, batching, and
other things we'd trip over when we iterate. Also removing analytics,
code related to performance. I'll come back to this soon and finish up.
3. New directory structure. Represents the work as we currently
understand it
