+++ static/app/routes.tsx
@@ -1689,24 +1689,54 @@ function buildRoutes() {
           {performanceChildRoutes}
         </Route>
       )}
-      {usingCustomerDomain && (
-        <Route
-          path="/starfish/"
-          component={withDomainRequired(
-            make(() => import('sentry/views/starfish/content'))
-          )}
-        />
-      )}
-      <Route
-        path="organizations/:orgId/starfish/"
-        component={make(() => import('sentry/views/starfish/content'))}
-      />
       <Route
         path="/organizations/:orgId/performance/"
         component={withDomainRedirect(make(() => import('sentry/views/performance')))}
         key="org-performance"
       >
         {performanceChildRoutes}
+      </Route>
+    </Fragment>
+  );
+
+  const starfishChildRoutes = (
+    <Fragment>
+      <IndexRoute
+        component={make(() => import('sentry/views/starfish/views/webServiceView'))}
+      />
+      <Route
+        path="database/"
+        component={make(() => import('sentry/views/starfish/modules/databaseModule'))}
+      />
+      <Route
+        path="api/"
+        component={make(() => import('sentry/views/starfish/modules/APIModule'))}
+      />
+      <Route
+        path="cache/"
+        component={make(() => import('sentry/views/starfish/modules/cacheModule'))}
+      />
+    </Fragment>
+  );
+
+  const starfishRoutes = (
+    <Fragment>
+      {usingCustomerDomain && (
+        <Route
+          path="/starfish/"
+          component={withDomainRequired(make(() => import('sentry/views/starfish')))}
+          key="orgless-starfish-route"
+        >
+          {starfishChildRoutes}
+        </Route>
+      )}
+
+      <Route
+        path="organizations/:orgId/starfish/"
+        component={make(() => import('sentry/views/starfish/views/webServiceView'))}
+        key="org-starfish"
+      >
+        {starfishChildRoutes}
       </Route>
     </Fragment>
   );
@@ -2137,6 +2167,7 @@ function buildRoutes() {
       {statsRoutes}
       {discoverRoutes}
       {performanceRoutes}
+      {starfishRoutes}
       {profilingRoutes}
       {adminManageRoutes}
       {gettingStartedRoutes}
