--- static/app/views/starfish/landing/widgets/transforms/transformHistogramQuery.tsx
@@ -1,12 +1,0 @@
-import {HistogramQueryChildrenProps} from 'sentry/utils/performance/histogram/histogramQuery';
-
-export function transformHistogramQuery(_: any, results: HistogramQueryChildrenProps) {
-  const {histograms} = results;
-  return {
-    ...results,
-    data: histograms,
-    isLoading: results.isLoading,
-    isErrored: results.error !== null,
-    hasData: !!Object.values(histograms || {}).length,
-  };
-}
