+++ static/app/views/starfish/modules/databaseModule/index.tsx
@@ -1,0 +1,29 @@
+import {Fragment} from 'react';
+
+import * as Layout from 'sentry/components/layouts/thirds';
+import {t} from 'sentry/locale';
+import {
+  PageErrorAlert,
+  PageErrorProvider,
+} from 'sentry/utils/performance/contexts/pageError';
+
+export default function DatabaseModule() {
+  return (
+    <Layout.Page>
+      <PageErrorProvider>
+        <Layout.Header>
+          <Layout.HeaderContent>
+            <Layout.Title>{t('Database')}</Layout.Title>
+          </Layout.HeaderContent>
+        </Layout.Header>
+
+        <Layout.Body>
+          <Layout.Main fullWidth>
+            <PageErrorAlert />
+            <Fragment />
+          </Layout.Main>
+        </Layout.Body>
+      </PageErrorProvider>
+    </Layout.Page>
+  );
+}
