--- static/app/views/starfish/types.tsx
@@ -1,5 +1,0 @@
-import EventView from 'sentry/utils/discover/eventView';
-
-import {QUERY_KEYS} from './utils';
-
-export type ViewProps = Pick<EventView, (typeof QUERY_KEYS)[number]>;
